Refactor CMSIS defines
