+++ system/MFL/MFL/Source/STARTUP/STARTUP.cpp
@@ -18,6 +18,7 @@
 //
 
 #include <stdint.h>
+#include <type_traits>
 
 #include "STARTUP.hpp"
 #include "FMC.hpp"
